bumped date function to utils
